Add flag to hide TOC header
This update should address
https://github.com/ekalinin/github-markdown-toc.go/issues/9.

I wanted to leave the default behavior to include a TOC header, but the
way https://github.com/alecthomas/kingpin#boolean-values work make it
difficult, so I head to create a negated flag.

I wanted to create a `header` flag, but I was not able to setup a
default value for a boolean flag. I was not able to discern a missing
flag from a negative value.

I thought of calling the flag `no-header`, but kingpin defines a
`--no-*` flag by default, so it would have ended up including a
`--no-no-header`, so I chose `--hide-header` instead.